Lag eget reguleringRepo
