+++ domain/src/main/kotlin/no/nav/su/se/bakover/domain/DatabaseRepos.kt
@@ -15,6 +15,7 @@ import no.nav.su.se.bakover.domain.oppdrag.avstemming.AvstemmingRepo
 import no.nav.su.se.bakover.domain.oppdrag.utbetaling.UtbetalingRepo
 import no.nav.su.se.bakover.domain.person.PersonRepo
 import no.nav.su.se.bakover.domain.personhendelse.PersonhendelseRepo
+import no.nav.su.se.bakover.domain.regulering.ReguleringRepo
 import no.nav.su.se.bakover.domain.revurdering.RevurderingRepo
 import no.nav.su.se.bakover.domain.sak.SakRepo
 import no.nav.su.se.bakover.domain.søknad.SøknadRepo
@@ -42,4 +43,5 @@ data class DatabaseRepos(
     val klageVedtakRepo: KlagevedtakRepo,
     val kontrollsamtaleRepo: KontrollsamtaleRepo,
     val avkortingsvarselRepo: AvkortingsvarselRepo,
+    val reguleringRepo: ReguleringRepo,
 )
