+++ web/src/main/kotlin/no/nav/su/se/bakover/web/routes/regulering/HentReguleringer.kt
@@ -8,10 +8,12 @@ import io.ktor.response.respond
 import io.ktor.routing.Route
 import io.ktor.routing.get
 import no.nav.su.se.bakover.common.serialize
+import no.nav.su.se.bakover.domain.Brukerrolle
 import no.nav.su.se.bakover.service.regulering.ReguleringService
 import no.nav.su.se.bakover.web.Resultat
 import no.nav.su.se.bakover.web.errorJson
 import no.nav.su.se.bakover.web.external.formatter
+import no.nav.su.se.bakover.web.features.authorize
 import no.nav.su.se.bakover.web.svar
 import java.time.Clock
 import java.time.LocalDate
@@ -33,21 +35,21 @@ internal fun Route.hentListeAvVedtakSomKanReguleres(
     // post kjørgreguleringautomatisk
     // post kjørdennemanuelt
 
-    // authorize(Brukerrolle.Saksbehandler) {
-    get("$reguleringPath") {
-        val dato = call.request.queryParameters["dato"].let {
-            if (it == null) {
-                LocalDate.of(LocalDate.now(clock).year, 5, 1)
-            } else {
-                formaterDato(it).getOrHandle {
-                    call.svar(it)
-                    return@get
+    authorize(Brukerrolle.Drift) {
+        get("$reguleringPath") {
+            val dato = call.request.queryParameters["dato"].let {
+                if (it == null) {
+                    LocalDate.of(LocalDate.now(clock).year, 5, 1)
+                } else {
+                    formaterDato(it).getOrHandle {
+                        call.svar(it)
+                        return@get
+                    }
                 }
             }
+            val aktiveBehandlinger =
+                reguleringService.hentAlleSakerSomKanReguleres(dato).getOrHandle { call.respond("feil") }
+            call.svar(Resultat.json(HttpStatusCode.Created, serialize(aktiveBehandlinger)))
         }
-        val aktiveBehandlinger =
-            reguleringService.hentAlleSakerSomKanReguleres(dato).getOrHandle { call.respond("feil") }
-        call.svar(Resultat.json(HttpStatusCode.Created, serialize(aktiveBehandlinger)))
     }
-    // }
 }
