+++ web/src/test/kotlin/no/nav/su/se/bakover/web/TestDatabaseBuilder.kt
@@ -15,6 +15,7 @@ import no.nav.su.se.bakover.domain.oppdrag.avstemming.AvstemmingRepo
 import no.nav.su.se.bakover.domain.oppdrag.utbetaling.UtbetalingRepo
 import no.nav.su.se.bakover.domain.person.PersonRepo
 import no.nav.su.se.bakover.domain.personhendelse.PersonhendelseRepo
+import no.nav.su.se.bakover.domain.regulering.ReguleringRepo
 import no.nav.su.se.bakover.domain.revurdering.RevurderingRepo
 import no.nav.su.se.bakover.domain.sak.SakRepo
 import no.nav.su.se.bakover.domain.søknad.SøknadRepo
@@ -45,6 +46,7 @@ object TestDatabaseBuilder {
         klageVedtakRepo: KlagevedtakRepo = mock(),
         kontrollsamtaleRepo: KontrollsamtaleRepo = mock(),
         avkortingsvarselRepo: AvkortingsvarselRepo = mock(),
+        reguleringRepo: ReguleringRepo = mock(),
     ): DatabaseRepos {
         return DatabaseRepos(
             avstemming = avstemming,
@@ -67,6 +69,7 @@ object TestDatabaseBuilder {
             klageVedtakRepo = klageVedtakRepo,
             kontrollsamtaleRepo = kontrollsamtaleRepo,
             avkortingsvarselRepo = avkortingsvarselRepo,
+            reguleringRepo = reguleringRepo,
         )
     }
 }
